Set zip_safe flag to False to fix reading data_files error
The `python setup.py install` command will install the package as an egg
by default. But data_files is not supported when installing packages as
eggs. Though data files in a package can be read using
`importlib.resources` module, the module was new in version 3.7. So here
we can set zip_safe flag to False to avoid the package installed as an
egg for the support of >= Python 3.6 .

Document of data_files:
https://packaging.python.org/en/latest/guides/distributing-packages-using-setuptools/#data-files

Document of importlib.resources module:
https://docs.python.org/3/library/importlib.html#module-importlib.resources

Document of setting the zig_safe flag:
https://setuptools.pypa.io/en/latest/userguide/miscellaneous.html#setting-the-zip-safe-flag

Other references:

https://stackoverflow.com/questions/6301003/stopping-setup-py-from-installing-as-egg

https://stackoverflow.com/questions/49782111/access-a-file-from-a-python-egg